Add files via upload
The archive skin_disease_50 contains the notebook file and the data organized in the correct folder structure to run the analysis on the data set containing 50 images per class, with a resolution of 128 by 128 pixels. The archive skin_disease_100 contains the relevant files for the dataset containing 100 images per class (with 2 exceptions). You can also find the notebook files separately, as well as some small python scripts for splitting the data into training and test (same as validation in this case) folders, for the VGG and Inception algorithms.